% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getData}
\alias{getData}
\title{Query SKM Exergy API.}
\usage{
getData(series_names, fromDate, toDate, fromHour = "00", toHour = "23")
}
\arguments{
\item{series_names}{Character vector with names of series. Example: c("CNPPLAN_RU1_H","PROPLAN_RU1_H","SPOT_RU1_H")}

\item{fromDate}{Start date as character (YYYY-MM-DD) or as date in R format. Examples:"2014-21-08" or "as.Date(Sys.time())"}

\item{toDate}{End date as character (YYYY-MM-DD) or as date in R format. Examples:"2014-22-08" or "as.Date(Sys.time())+1"}

\item{fromHour}{Start hour as character ("HH"). "00" by default for first hour of day.}

\item{toHour}{End hour as character ("HH"). "23" by default for the last hour of day.}
}
\value{
Dataframe with data for series and time period.
}
\description{
This function creates a web query url and reads data from SKM Exergy (\url{https://exergy.skmenergy.com/}) into a data.frame. You must be a client for use it.
}
\examples{
\dontrun{
# Get data for the last week
today<-as.Date(Sys.time())
seriesList<-c("CNPPLAN_RU1_H","PROPLAN_RU1_H","SPOT_RU1_H")

dataSet<-getData(seriesList,today-6,today)
dataSet}
}

